requirements.txt
rembg
scikit-learn
uvicorn
python-multipart
tensorflow
pixellib
annotated-types
anyio
certifi
charset-normalizer
click
colorama
cycler
exceptiongroup
fastapi
fonttools
h11
idna
importlib-metadata
Jinja2
kiwisolver
MarkupSafe
matplotlib
numpy
opencv-python
packaging
Pillow
pydantic
pydantic_core
pyparsing
python-dateutil
requests
six
sniffio
starlette
torch
torchvision
typing_extensions
urllib3
zipp
PyJWT
python-decouple==3.3